requirements.txt
cohere
streamlit
python-dotenv
streamlit-oauth